+++ liana/method/sp/_SpatialBivariate.py
@@ -6,6 +6,7 @@ from scipy.sparse import csr_matrix
 from anndata import AnnData
 from mudata import MuData
 from liana.method._pipe_utils._common import _get_props
+from liana._constants._docs import d
 
 from liana.method.sp._spatial_pipe import (
     _categorize,
@@ -28,8 +29,9 @@ from liana._logging import _logg
 
 class SpatialBivariate():
     """ A class for bivariate local spatial metrics. """
-    def __init__(self):
-        pass
+    def __init__(self, x_name='x', y_name='y'):
+        self.x_name = x_name
+        self.y_name = y_name
     
     def _handle_return(self, data, stats, local_scores, key_added, x_added, inplace=False):
         if not inplace:
@@ -42,37 +44,34 @@ class SpatialBivariate():
         else:
             data.obsm[x_added] = local_scores
         
-
+    @d.dedent
     def __call__(self,
-                 mdata,
-                 x_mod,
-                 y_mod,
-                 function_name='cosine',
-                 interactions = None,
-                 resource=None,
-                 resource_name=None,
-                 connectivity_key = 'spatial_connectivities',
-                 mod_added = "local_scores",
-                 key_added = 'global_res',
-                 mask_negatives=False,
-                 add_categories = False,
+                 mdata: (MuData | AnnData),
+                 x_mod: str,
+                 y_mod: str,
+                 function_name: str='cosine',
+                 interactions: (None | list) = None,
+                 resource: (None | pd.DataFrame) = None,
+                 resource_name: (None | str) = None,
+                 connectivity_key:str = 'spatial_connectivities',
+                 mod_added:str = "local_scores",
+                 key_added:str = 'global_res',
+                 mask_negatives:bool=False,
+                 add_categories:bool = False,
                  n_perms: int = None,
-                 seed = 1337,
-                 nz_threshold = 0, # NOTE: do I rename this?
-                 x_use_raw = False,
-                 x_layer = None,
-                 x_transform = False,
-                 y_use_raw=False,
-                 y_layer = None,
-                 y_transform = False,
-                 # TODO: move some of these to self
-                 x_name='x',
-                 y_name='y',
-                 complex_sep=None,
-                 xy_sep = '^',
-                 remove_self_interactions=True,
-                 inplace = True,
-                 verbose=False,
+                 seed:int = 1337,
+                 nz_threshold:float = 0, # NOTE: do I rename this?
+                 x_use_raw:bool = False,
+                 x_layer: (None | str) = None,
+                 x_transform: (bool | callable) = False,
+                 y_use_raw:bool = False,
+                 y_layer: (None | str) = None,
+                 y_transform: (bool | callable) = False,
+                 complex_sep: (None | str)=None,
+                 xy_sep:str = '^',
+                 remove_self_interactions:bool=True,
+                 inplace:bool = True,
+                 verbose:bool = False,
                  ):
         """
         A method for bivariate local spatial metrics.
@@ -80,35 +79,21 @@ class SpatialBivariate():
         Parameters
         ----------
         
-        mdata: MuData or AnnData
-            MuData or AnnData object with spatial coordinates.
-        x_mod: str
-            Name of the modality to use for the x-axis.
-        y_mod: str
-            Name of the modality to use for the y-axis.
-        function_name: str
-            Name of the function to use for the analysis.
-        interactions: list
-            List of tuples with ligand-receptor pairs `[(ligand, receptor), ...]` to be used for the analysis.
-            If passed, it will overrule the resource requested via `resource` and `resource_name`.
-        resource: pd.DataFrame
-            Resource to use for the analysis. If None, `resource_name` is used.
-        resource_name: str
-            Name of the resource to use for the analysis. If None, the default resource ('consensus') is used.
-        connectivity_key: str
-            Key in `mdata.uns` where the spatial connectivities are stored.
+        %(mdata)s
+        %(x_mod)s
+        %(y_mod)s
+        %(function_name)s
+        %(interactions)s
+        %(resource)s
+        %(resource_name)s
+        %(connectivity_key)s
         mod_added: str
             Key in `mdata.mod` where the local scores are stored.
-        key_added: str
-            Key in `mdata.uns` where the global scores are stored.
-        mask_negatives: bool
-            Whether to mask negative-negative (low-low) or uncategorized interactions.
-        add_categories: bool
-            Whether to add categories about the local scores
-        n_perms: int
-            Number of permutations to use for the analysis. If None, no p-values are computed. If 0 analytical p-values are computed.
-        seed: int
-            Seed for the random number generator.
+        %(key_added)s
+        %(mask_negatives)s
+        %(add_categories)s
+        %(n_perms)s
+        %(seed)s
         nz_threshold: float
             Minimum proportion of cells expressing the ligand and receptor.
         x_use_raw: bool
@@ -123,20 +108,14 @@ class SpatialBivariate():
             Layer to use for y-mod.
         y_transform: bool
             Function to transform the y-mod.
-        x_name: str
-            Name of the x-mod.
-        y_name: str
-            Name of the y-mod.
         complex_sep: str
             Separator to use for complex names.
         xy_sep: str
             Separator to use for interaction names.
         remove_self_interactions: bool
             Whether to remove self-interactions. `True` by default.
-        inplace: bool
-            Whether to add the results to `mdata` or return them.
-        verbose: bool
-            Verbosity flag.
+        %(inplace)s
+        %(verbose)s
         
         Returns
         -------
@@ -144,7 +123,6 @@ class SpatialBivariate():
         If `inplace` is `True`, the results are added to `mdata` and `None` is returned.
         If `inplace` is `False`, the results are returned.
         """
-        
         
         if n_perms is not None:
             if not isinstance(n_perms, int) or n_perms < 0:
@@ -155,8 +133,8 @@ class SpatialBivariate():
         resource = _handle_resource(interactions=interactions,
                                     resource=resource,
                                     resource_name=resource_name,
-                                    x_name=x_name,
-                                    y_name=y_name,
+                                    x_name=self.x_name,
+                                    y_name=self.y_name,
                                     verbose=verbose)
         
         if isinstance(mdata, MuData):
@@ -195,25 +173,25 @@ class SpatialBivariate():
         
         if complex_sep is not None:
             adata = _add_complexes_to_var(adata,
-                                          np.union1d(resource[x_name].astype(str),
-                                                     resource[y_name].astype(str)
+                                          np.union1d(resource[self.x_name].astype(str),
+                                                     resource[self.y_name].astype(str)
                                                      ),
                                           complex_sep=complex_sep
                                           )
         
         # filter_resource
-        resource = resource[(np.isin(resource[x_name], adata.var_names)) &
-                            (np.isin(resource[y_name], adata.var_names))]
+        resource = resource[(np.isin(resource[self.x_name], adata.var_names)) &
+                            (np.isin(resource[self.y_name], adata.var_names))]
         
         # NOTE: Should I just get rid of remove_self_interactions?
-        self_interactions = resource[x_name] == resource[y_name]
+        self_interactions = resource[self.x_name] == resource[self.y_name]
         if self_interactions.any() & remove_self_interactions:
             _logg(f"Removing {self_interactions.sum()} self-interactions", verbose=verbose)
             resource = resource[~self_interactions]
 
         # get entities
-        entities = np.union1d(np.unique(resource[x_name]),
-                                np.unique(resource[y_name]))
+        entities = np.union1d(np.unique(resource[self.x_name]),
+                                np.unique(resource[self.y_name]))
         # Check overlap between resource and adata TODO check if this works
         assert_covered(entities, adata.var_names, verbose=verbose)
 
@@ -225,18 +203,18 @@ class SpatialBivariate():
                                 index=adata.var_names
                                 ).reset_index().rename(columns={'index': 'gene'})
         # join global stats to LRs from resource
-        xy_stats = resource.merge(_rename_means(xy_stats, entity=x_name)).merge(
-            _rename_means(xy_stats, entity=y_name))
+        xy_stats = resource.merge(_rename_means(xy_stats, entity=self.x_name)).merge(
+            _rename_means(xy_stats, entity=self.y_name))
         
         # TODO: nz_threshold to nz_prop? For consistency with other methods
         # filter according to props
-        xy_stats = xy_stats[(xy_stats[f'{x_name}_props'] >= nz_threshold) &
-                            (xy_stats[f'{y_name}_props'] >= nz_threshold)]
+        xy_stats = xy_stats[(xy_stats[f'{self.x_name}_props'] >= nz_threshold) &
+                            (xy_stats[f'{self.y_name}_props'] >= nz_threshold)]
         # create interaction column
-        xy_stats['interaction'] = xy_stats[x_name] + xy_sep + xy_stats[y_name]
+        xy_stats['interaction'] = xy_stats[self.x_name] + xy_sep + xy_stats[self.y_name]
         
-        x_mat = adata[:, xy_stats[x_name]].X.T
-        y_mat = adata[:, xy_stats[y_name]].X.T
+        x_mat = adata[:, xy_stats[self.x_name]].X.T
+        y_mat = adata[:, xy_stats[self.y_name]].X.T
         
         # reorder columns, NOTE: why?
         xy_stats = xy_stats.reindex(columns=sorted(xy_stats.columns))