+++ liana/method/sp/_lr_bivar.py
@@ -5,13 +5,14 @@ import pandas as pd
 from typing import Optional
 
 from liana.method.sp._SpatialBivariate import SpatialBivariate
-
+from liana._constants._docs import d
 
 class SpatialLR(SpatialBivariate):
     """ A child class of SpatialBivariate for ligand-receptor analysis. """
     def __init__(self):
-        super().__init__()
+        super().__init__(x_name='ligand', y_name='receptor')
 
+    @d.dedent
     def __call__(self,
                  adata: AnnData,
                  function_name: str,
@@ -19,7 +20,7 @@ class SpatialLR(SpatialBivariate):
                  resource: Optional[pd.DataFrame] = None,
                  interactions=None,
                  expr_prop: float = 0.05,
-                 n_perms: int = None,
+                 n_perms:(None | int) = None,
                  mask_negatives: bool = False,
                  seed: int = 1337,
                  add_categories: bool = False,
@@ -37,43 +38,25 @@ class SpatialLR(SpatialBivariate):
         
         Parameters
         ----------
-        adata: AnnData
-            AnnData object with spatial coordinates.
-        function_name: str
-            Name of the function to use for the analysis.
-        resource_name: str
-            Name of the resource to use for the analysis.
-        resource: pd.DataFrame
-            Resource to use for the analysis. If None, the default resource ('consensus') is used.
-        interactions: list
-            List of tuples with ligand-receptor pairs `[(ligand, receptor), ...]` to be used for the analysis.
-            If passed, it will overrule the resource requested via `resource` and `resource_name`.
-        expr_prop: float
-            Minimum proportion of cells expressing the ligand and receptor.
-        n_perms: int
-            Number of permutations to use for the analysis.
-        positive_only: bool
-            Whether to mask non-positive interactions.
-        seed: int
-            Seed for the random number generator.
-        add_categories: bool
-            Whether to add categories about the local scores
-        use_raw: bool
-            Whether to use .raw attribute of adata. If False, .X is used.
-        layer: str
-            Name of the layer to use. If None, .X is used.
-        connectivity_key: str
-            Key in `adata.obsp` where the spatial connectivities are stored.
-        inplace: bool
-            Whether to add the results to `adata.uns` and `adata.obsm` or return them.
-        key_added: str
-            Key in `adata.uns` where the global results are stored.
+        %(adata)s
+        %(function_name)s
+        %(resource_name)s
+        %(resource)s
+        %(interactions)s
+        %(expr_prop)s
+        %(n_perms)s
+        %(mask_negatives)s
+        %(seed)s
+        %(add_categories)s
+        %(use_raw)s
+        %(layer)s
+        %(connectivity_key)s
+        %(inplace)s
+        %(key_added)s
         obsm_added: str
             Key in `adata.obsm` where the local scores are stored.
-        lr_sep: str
-            Separator to use between ligand and receptor names.
-        verbose: bool
-            Whether to print progress messages.
+        %(lr_sep)s
+        %(verbose)s
         
         Returns
         -------
@@ -101,8 +84,6 @@ class SpatialLR(SpatialBivariate):
             seed=seed,
             verbose=verbose,
             xy_sep=lr_sep,
-            x_name='ligand',
-            y_name='receptor',
             inplace=False,
             complex_sep='_'
             )